requirements.txt
pandas
numpy
scikit-learn
python-box
pydantic
PyYAML
logging
requests
zipfile36
xgboost
imbalanced-learn
imbalance
joblib
matplotlib
seaborn
plotly
shap
catboost
uvicorn[standard]
fastapi